LayoutTests/fast/selectors/querySelector-long-multiple-attribute-match-with-child-backtracking-expected.txt

Test backtracking of matching multiple attribute with multiple child selector relations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("[data-foo][data-bar]>[data-webkit][data-padding]>div>pre span[style]").length is 3
PASS document.querySelectorAll("[data-foo][data-bar]>[data-webkit][data-padding]>div>pre span[style]")[0].id is "target1"
PASS document.querySelectorAll("[data-foo][data-bar]>[data-webkit][data-padding]>div>pre span[style]")[1].id is "target2"
PASS document.querySelectorAll("[data-foo][data-bar]>[data-webkit][data-padding]>div>pre span[style]")[2].id is "target3"
PASS successfullyParsed is true

TEST COMPLETE
